requirements.txt
discord.py
requests
python-dotenv
Flask
psutil==5.8.0
